Extracted classpath methods
